ml/api/requirements.txt
# Flask API Dependencies
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.10.0
Pillow>=9.0.0
numpy>=1.23.0
gunicorn>=21.0.0  # For production deployment
